LAG-14: moved help button into items component for the format
